citation fields & unit tests
